Enable verification of token database
When verification fails, the user either needs to start
with -reindex-tokens manually, or let the wallet reindex itself.
Around the upgrade to 5.0, a full reindex is a sane default.

- Rename -reindextokens to -reindex-tokens
- Move ReindexTokenDB after setting chainActive
- Add VerifyTokenDB
- Add block hash of token creation to db
- Do not buffer tokens when reindexing

+++ src/tokens/rpctokenwallet.cpp
@@ -965,7 +965,7 @@ extern UniValue configuretokendryrun(const JSONRPCRequest& request)
     CTokenGroupDescription tokenGroupDescription(opretScript);
     CTokenGroupStatus tokenGroupStatus;
     CTransaction dummyTransaction;
-    CTokenGroupCreation tokenGroupCreation(MakeTransactionRef(dummyTransaction), tokenGroupInfo, tokenGroupDescription, tokenGroupStatus);
+    CTokenGroupCreation tokenGroupCreation(MakeTransactionRef(dummyTransaction), uint256(), tokenGroupInfo, tokenGroupDescription, tokenGroupStatus);
     tokenGroupCreation.ValidateDescription();
 
     ret.push_back(Pair("ticker", tokenGroupCreation.tokenGroupDescription.strTicker));
